pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "promptlab"
version = "0.1.0"
description = "MVP for testing LLM portfolio allocation under V3 rules"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.24.0",
    "dashscope>=1.20.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
]

[project.scripts]
promptlab = "promptlab.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/promptlab"]
